did more cleaning and added a registration page
