+++ our-app/src/App.css
@@ -15,7 +15,7 @@
 
 .App-header {
   background-color: #282c34;
-  min-height: 100vh;
+  min-height: 10vh;
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -26,6 +26,19 @@
 
 .App-link {
   color: #61dafb;
+}
+ form label {
+   padding-top: 3%;
+   margin-left:25%;
+   width: 50%;
+}
+form label .register-input {
+  border: black solid 2px;
+}
+.form-buttons {
+  background-color: grey;
+  color:white;
+  margin-bottom: 2%;
 }
 
 @keyframes App-logo-spin {
