--- our-app/src/Register.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-
-function Register () {
-        <div>
-            <form>
-                <label>
-                    
-                </label>
-            </form>
-        </div>
-}
-
-export default Register
